ejercicio8-distribuidora/src/main/java/ar/edu/unlp/info/oo1/ej8_distribuidora/Factura.java
package ar.edu.unlp.info.oo1.ej8_distribuidora;

import java.time.LocalDate;

public class Factura {
	private double montoEnergiaActiva;
	private double descuento;
	private LocalDate fecha;
	private Usuario usuario;

	public Factura(double montoEnergiaActiva, double descuento, Usuario usuario) {
		this.montoEnergiaActiva = montoEnergiaActiva;
		this.descuento = descuento;
		this.fecha = LocalDate.now();
		this.usuario = usuario;
	}

	public double getMontoEnergiaActiva() {
		return montoEnergiaActiva;
	}

	public double getDescuento() {
		return descuento;
	}

	public LocalDate getFecha() {
		return fecha;
	}

	public Usuario getUsuario() {
		return usuario;
	}

	public double montoTotal() {
		return  this.montoEnergiaActiva * (1 - this.descuento / 100);
	}

}
